launch.sh: change inventories dir in docker
The CI is run in a Docker container and should access the local
inventories and keys.
This is done by mapping /home/virtu/ansible on /tmp inside docker.

However, other softwares may be using /tmp and conflict may occur. It is
thus a better practice to map it on a sub dir inside /tmp, here /tmp/ci.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>